added associated section warning and renamed old components
