+++ frontend/src/App.js
@@ -3,7 +3,6 @@ import 'bulma/css/bulma.css'
 import './styles/App.css';
 import Sections from "./components/selector/Sections";
 import Schedule from './components/schedule/Schedule'
-import ScheduleNew from './components/schedule/ScheduleNew'
 
 import Provider from "react-redux/es/components/Provider";
 import {applyMiddleware, createStore} from "redux";
@@ -83,7 +82,7 @@ class App extends Component {
                             </div>
                         </div>
                         <div id="InfoPanel" className="column">
-                            <ScheduleNew/>
+                            <Schedule/>
                         </div>
                     </div>
                 </div>
