+++ frontend/src/components/schedule/schedule.css
@@ -50,6 +50,17 @@
     opacity: 1;
 }
 
+.hide {
+    display: none;
+}
+
+.warning {
+    position: absolute;
+    top: 3px;
+    right: 4px;
+    font-size: 1em;
+}
+
 .time {
     position: absolute;
     top: -6px;
